Profile section full integration
Added the ProfileSection entity that represents section data in a profile, and added the ability to prefill it to a resume, create and change them, etc... (Basically, all abilities that were already present for the other profile and resume entities)

+++ src/main/java/com/rebuild/backend/controllers/ProfileController.java
@@ -3,10 +3,7 @@ package com.rebuild.backend.controllers;
 import com.rebuild.backend.exceptions.profile_exceptions.NoProfileException;
 import com.rebuild.backend.model.entities.users.User;
 import com.rebuild.backend.model.entities.profile_entities.UserProfile;
-import com.rebuild.backend.model.forms.profile_forms.FullProfileForm;
-import com.rebuild.backend.model.forms.profile_forms.ProfileEducationForm;
-import com.rebuild.backend.model.forms.profile_forms.ProfileExperienceForm;
-import com.rebuild.backend.model.forms.profile_forms.ProfileHeaderForm;
+import com.rebuild.backend.model.forms.profile_forms.*;
 import com.rebuild.backend.service.user_services.ProfileService;
 import com.rebuild.backend.service.user_services.UserService;
 import jakarta.validation.Valid;
@@ -73,39 +70,86 @@ public class ProfileController {
         return profileService.updateProfileExperiences(profile, experienceFormList);
     }
 
+    @PatchMapping("/patch/sections")
+    @ResponseStatus(HttpStatus.OK)
+    public UserProfile updateProfileSections(@Valid @RequestBody List<ProfileSectionForm> sectionFormList,
+                                             @AuthenticationPrincipal User authenticatedUser){
+        if(authenticatedUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
+        UserProfile profile = authenticatedUser.getProfile();
+        return profileService.updateProfileSections(profile, sectionFormList);
+    }
+
     @DeleteMapping("/delete")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteProfile(@AuthenticationPrincipal User deletingUser){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
         profileService.deleteProfile(deletingUser.getProfile().getId());
     }
 
     @DeleteMapping("/delete/header")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteProfileHeader(@AuthenticationPrincipal User deletingUser){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
         profileService.deleteProfileHeader(deletingUser.getProfile().getId());
     }
 
     @DeleteMapping("/delete/education")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteProfileEducation(@AuthenticationPrincipal User deletingUser){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
         profileService.deleteProfileEducation(deletingUser.getProfile().getId());
     }
 
     @DeleteMapping("/delete/experiences")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteProfileExperience(@AuthenticationPrincipal User deletingUser){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
         profileService.deleteProfileExperiences(deletingUser.getProfile().getId());
     }
 
     @DeleteMapping("/delete/experiences/{experience_id}")
-    @ResponseStatus(HttpStatus.OK)
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public UserProfile deleteSpecificExperience(@AuthenticationPrincipal User deletingUser,
                                                 @PathVariable UUID experience_id){
-        return profileService.deleteSpecificProfileExperience(deletingUser.getProfile().getId(),
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
+        return profileService.deleteSpecificProfileExperience(deletingUser.getProfile(),
                 experience_id);
     }
 
+    @DeleteMapping("/delete/sections")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteProfileSections(@AuthenticationPrincipal User deletingUser){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
+        profileService.deleteProfileSections(deletingUser.getProfile().getId());
+    }
+
+    @DeleteMapping("/delete/sections/{section_id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public UserProfile deleteSpecificSection(@AuthenticationPrincipal User deletingUser,
+                                                @PathVariable UUID section_id){
+        if(deletingUser.getProfile() == null){
+            throw new NoProfileException("No profile found for your account");
+        }
+        return profileService.deleteSpecificSection(deletingUser.getProfile(),
+                section_id);
+    }
+
     @PutMapping("/update_profile")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public UserProfile changeEntireProfile(@AuthenticationPrincipal User updatingUser,
                                            @Valid @RequestBody FullProfileForm profileForm){
         return profileService.updateEntireProfile(updatingUser.getProfile(), profileForm);
